Save/Load working, added 'save-reset' div, refactored saveToStorage/updateFromStorage
Save and load is functional, although no option to choose username.
added an area to the layout for the save/reset/load buttons within the
hint area. made this scroll if it overflows/keep save-reset div fixed to
bottom. had to refactor saveToStorage/updateFromStorage in order to call
from links and be able to save under different usernames. can load from
other usernames but no way for user to accomplish on page.