Keys disabled with pop up
Had to refactor initlesson, and removepopup.. but keys are now disabled
while their is a pop up open. This is done by running killkeys() of
current lesson when pop up opens, and then runnign initlesson() when pop
up closes. It only attempts to disable keys if there is text in vim text
area.