Multiple finish Cond working with column position
by setting the value to an array, you can specify multiple values that
will satisfy the finish condition. this is accomplished through the
logic of: if finish condition is array or object & this key is not in
lesson then see if value in lesson .includes() in finishcond. else do
normal comparison.

this also includes the dark mode changes...