added j and k functionality, row/col logic
Added functionality of j and k keys. 3rd lesson. the current row column
logic doesn't quite follow vim.. in true vim, when you change rows from a row
that is longer and move back, it moves back to the end of that row. some
how like remembering 'furthestCol' in a way. current way does not break,
but mimicking true vim will be needed.